requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
cloudinary==1.44.1
cryptography==45.0.5
dnspython==2.7.0
Faker==37.4.2
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
julseb-lib-python-utils==0.1.0
jwt==1.4.0
MarkupSafe==3.0.2
mongoengine==0.29.1
packaging==25.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.13.2
python-dotenv==1.1.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
